+++ app/mypage/page.tsx
@@ -1,19 +1,20 @@
 'use client';
 
-import { useSearchParams } from 'next/navigation';
-import { useRouter } from 'next/navigation';
+import { use } from 'react';
+
+import { useRouter, useSearchParams } from 'next/navigation';
 
 import ManageMembership from './static/MyProfileMemberships';
-import UserDataOverview from './static/MyProfileBasics';
+import MyProfileBasics from './static/MyProfileBasics';
+
 import { apiMyPageBasics } from '@/lib/api/calls/mypage';
+import { CloseIcon } from '@/component/icon';
+import { LoginRequiredError } from '@/lib/api/errors';
 
 import TempAlert from './static/TempAlert';
-import { CloseIcon } from '@/lib/icons';
 
-import { use } from 'react';
-import BookmarksView from './dynamic/BookmarksView';
-import MyCommentsView from './dynamic/MyCommentsView';
-import { LoginRequiredError } from '@/lib/api/errors';
+import BookmarksView from './part/MyBookmarksView';
+import MyCommentsView from './part/MyCommentsView';
 
 export default function ProfileOverviewWithMemberShip() {
   const router = useRouter();
@@ -42,7 +43,7 @@ export default function ProfileOverviewWithMemberShip() {
     <div className="flex flex-col w-full h-full">
       {pwchanged ? <TempAlert closeButton={RemoveAlert}>비밀번호가 성공적으로 변경되었습니다.</TempAlert> : <></>}
       {profilechanged ? <TempAlert closeButton={RemoveAlert}>프로필을 성공적으로 업데이트했습니다.</TempAlert> : <></>}
-      <UserDataOverview userprofile={myprofile.data} />
+      <MyProfileBasics userprofile={myprofile.data} />
       <ManageMembership access_expiration_date={myprofile.data.access_expiration_date} />
       <BookmarksView />
       <MyCommentsView />
