+++ component/commonLayout/carousel/carousel.tsx
@@ -2,7 +2,7 @@
 
 import React, { useEffect, useState } from 'react';
 
-import { HStack } from '@/components/basic/stack';
+import { HStack } from '@/component/basic/stack';
 
 interface CarouselProps {
   children: React.ReactNode[];
