+++ component/commonLayout/topnav/aux/static.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link';
-import { LogoIconRich, UpIcon, DownIcon } from '@/lib/icons';
-import Skeleton from '@/components/composite/skeleton';
-import Button from '@/components/basic/button';
+import { LogoIconRich, UpIcon, DownIcon } from '@/component/icon';
+import Skeleton from '@/component/composite/skeleton';
+import Button from '@/component/basic/button';
 
 export function TopNavInnerLeft() {
   return (
